Make caret position into individual class.
