+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Algorithms/TimeTagCaretPositionAlgorithm.cs
@@ -5,6 +5,7 @@ using System;
 using System.Linq;
 using osu.Framework.Extensions.IEnumerableExtensions;
 using osu.Framework.Graphics.Sprites;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
 using osu.Game.Rulesets.Karaoke.Extensions;
 using osu.Game.Rulesets.Karaoke.Objects;
