+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/CaretPosition/ICaretPosition.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Game.Rulesets.Karaoke.Objects;
+
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition
+{
+    public interface ICaretPosition
+    {
+        public Lyric Lyric { get; }
+    }
+}
