+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Components/Lyrics/Carets/IHasCaretPosition.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
+
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Carets
 {
     public interface IHasCaretPosition
