+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/ILyricEditorState.cs
@@ -2,6 +2,7 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using osu.Framework.Bindables;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
 {
